Added a lot of fixes see the "CinchV1 Release Notes.docx" file which contains extensive notes about the fixes and breaking changes. This covers all issues/discussion points that I could fix.
NOTE : I am now commencing work on CinchV2 for .NET4/VS2010, and that will be quite different in terms of how the attached properties and services work the core ideas will remain the same.

I will be updating the existing Cinch articles over at codeproject with these changes, but any new Cinch work will only be done in CinchV2 (as it is the future of Cinch) unless someone reports an absolute deal breaker.

Anyway see the release "CinchV1 Release Notes.docx" for all that is new

git-svn-id: https://cinch.svn.codeplex.com/svn@46013 ffd33b8c-2492-42e0-bdc5-587b920b7d6d