Added PartCreationPolicy MEF attributes to ViewModels in WPF/SL demo projects (as it was confusing to some users to always be getting the same ViewModels (although that is ok for the demos, its not very real world like, so showed a more real world scenario, where each View would get a new instance of a ViewModel which is achieved using [PartCreationPolicy(CreationPolicy.NonShared)].
Also to enable this, I have included a specialised TabControl for WPF which allows all open Views to be stored in memory and just shows/hides active one. The standard TabControl in WPF works different to this where is creates entire VisualTree again, so that would be a disaster using MefedMVVM as that would mean we would get a new VM each time we change tabs. So the WPF demo shows you how to work around this using a special TabControlEx which has a code file and Style (AppStyles.xaml) that you can use. See the WPF demo app for that. 

As for SL, as the TabControl in SL lacks the correct overrides, I am afraid to say you are on your own in trying to figure out that one.


git-svn-id: https://cinch.svn.codeplex.com/svn@48252 ffd33b8c-2492-42e0-bdc5-587b920b7d6d