Adding enhancements to the WPF message box service based on Cinch user Michel Renauds comments. In fact Michel provided the work I just reviewed it and added his extra methods.
In his words this is why they needed adding

Early on when we started using Cinch, we noticed some limitations in the message box service, namely hardcoded captions (we sometimes need something more explicit, and the apps are bilingual too) and inability to set a specific button as the default button (e.g. having "Cancel" as the default instead of "Ok" ). I ended up adding more methods to the service to take care of that.

git-svn-id: https://cinch.svn.codeplex.com/svn@68724 ffd33b8c-2492-42e0-bdc5-587b920b7d6d